move cache to utils
