+++ pkg/security/utils/cache/lru_2layers.go
@@ -3,10 +3,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build linux
-
-// Package dentry holds dentry related files
-package dentry
+// Package cache holds cache related files
+package cache
 
 import (
 	"sync"
